requirements.txt
bayesmark==0.0.8
botorch==0.14.0
certifi==2025.6.15
charset-normalizer==3.4.2
cma==4.2.0
coco-experiment==2.8.1
cocopp==2.7.4
colorama==0.4.6
ConfigSpace==1.2.1
contourpy==1.3.2
cycler==0.12.1
filelock==3.18.0
fonttools==4.58.4
fsspec==2025.5.1
gitdb==4.0.12
GitPython==3.1.44
gpytorch==1.14
idna==3.10
importlib_metadata==8.7.0
jaxtyping==0.3.2
Jinja2==3.1.6
joblib==1.5.1
kiwisolver==1.4.8
linear-operator==0.6
MarkupSafe==3.0.2
matplotlib==3.10.3
more-itertools==10.7.0
mpmath==1.3.0
multipledispatch==1.0.0
mypy_extensions==1.1.0
networkx==3.5
numpy==2.3.0
opt_einsum==3.4.0
packaging==25.0
pandas==2.3.0
pathvalidate==3.3.1
pfns4bo==0.1.5
pillow==11.2.1
platformdirs==4.3.8
pyparsing==3.2.3
pyre-extensions==0.0.32
pyro-api==0.1.2
pyro-ppl==1.9.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
scikit-learn==1.1.3
scipy==1.15.3
six==1.17.0
smmap==5.0.2
sympy==1.14.0
threadpoolctl==3.6.0
torch==2.7.1+cu128
tqdm==4.67.1
typing-inspect==0.9.0
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
wadler_lindig==0.1.6
xarray==2025.6.1
zipp==3.23.0
